requirements.txt
psycopg2==2.9.6
pycryptodome==3.18.0
